split layout and options
