+++ src/components.d.ts
@@ -19,7 +19,7 @@ import { Input } from "./components/input/Input";
 import { RGB } from "./model/Color/RGB";
 import { RadioItemSelect } from "./components/input/radio/RadioItemSelect";
 import { SmoothlyInputRadio } from "./components/input/radio/index";
-import { MenuHeight } from "./components/input/select/options";
+import { layout } from "./components/input/select/layout";
 import { SmoothlyTabs } from "./components/tabs";
 export { Color, Data, Fill, Icon, Message, Notice, Submit, Trigger } from "./model";
 export { FunctionalComponent, VNode } from "@stencil/core";
@@ -35,7 +35,7 @@ export { Input } from "./components/input/Input";
 export { RGB } from "./model/Color/RGB";
 export { RadioItemSelect } from "./components/input/radio/RadioItemSelect";
 export { SmoothlyInputRadio } from "./components/input/radio/index";
-export { MenuHeight } from "./components/input/select/options";
+export { layout } from "./components/input/select/layout";
 export { SmoothlyTabs } from "./components/tabs";
 export namespace Components {
     interface SmoothlyApp {
@@ -786,7 +786,7 @@ export namespace Components {
         "invalid"?: boolean;
         "listen": (listener: Editable.Observer.Listener) => Promise<void>;
         "looks"?: Looks;
-        "menuHeight"?: MenuHeight;
+        "menuHeight"?: layout.MenuHeight;
         /**
           * @default false
          */
@@ -3325,7 +3325,7 @@ declare namespace LocalJSX {
          */
         "invalid"?: boolean;
         "looks"?: Looks;
-        "menuHeight"?: MenuHeight;
+        "menuHeight"?: layout.MenuHeight;
         /**
           * @default false
          */
@@ -3998,7 +3998,7 @@ declare namespace LocalJSX {
         "clearable": boolean;
         "defined": boolean;
         "placeholder": string;
-        "menuHeight": MenuHeight;
+        "menuHeight": layout.MenuHeight;
         "required": boolean;
         "searchDisabled": boolean;
         "mutable": boolean;
